fix: out-of-bounds access and unsigned wraparound in ListArray kernels
`awkward_ListArray_min_range` read `fromstops[0] - fromstarts[0]` and wrote
`*tomin` before checking `lenstarts`, so an empty ListArray read two empty
buffers and overwrote the caller's output. The specification guards the whole
body on `lenstarts > 0` and the CUDA path early-returns, so the CPU kernel was
the outlier.

Six kernels computed `int64_t length = fromstops[i] - fromstarts[i]`, where
the subtraction happens in the operands' own width before the widening. For
the `uint32` specializations a malformed `stops < starts` wrapped to ~4.29e9
instead of going negative, so `getitem_next_at` returned a bogus carry where
the `int32` specialization correctly reported `index out of range`; in the
range kernels the wrapped length reached `awkward_regularize_rangeslice` and
produced an unbounded write range rather than an empty result. Widen before
subtracting.

+++ awkward-cpp/src/cpu-kernels/awkward_ListArray_rpad_and_clip_length_axis1.cpp
@@ -13,7 +13,7 @@ ERROR awkward_ListArray_rpad_and_clip_length_axis1(
   int64_t lenstarts) {
   int64_t length = 0;
   for (int64_t i = 0;  i < lenstarts;  i++) {
-    int64_t rangeval = fromstops[i] - fromstarts[i];
+    int64_t rangeval = (int64_t)fromstops[i] - (int64_t)fromstarts[i];
     length += (target > rangeval) ? target : rangeval;
   }
   *tomin = length;
